requirements.txt
streamlit
python-dotenv
google-generativeai
PyPDF2
python-docx
fpdf2
